nimble/ll/dtm: Start RX with interrupts disabled
It can happen that if we are interrupted when setting things for RX,
we will be stuck with radio unconfigured. For now let's just make sure
this is done with interrupts disabled.

It may be reasonable to handle this also somehow in PHY, but it needs
more analysis and we don't want to break something. Also note that in
other cases we set wfr so even if this happens, wfr will rescue us.